# Faster bootstrap when using local agent.