--- a/cloudconfig/cloudinit/interface.go
+++ b/cloudconfig/cloudinit/interface.go
@@ -319,6 +319,10 @@ type WrittenFilesConfig interface {
 	// of a given file with the specified file permissions on *first* boot.
 	// NOTE: if the file already exists, it will be truncated.
 	AddRunBinaryFile(string, []byte, uint)
+
+	// SetFileTransporter sets an alternative transporter for files added currently
+	// with AddRunBinaryFile.
+	SetFileTransporter(FileTransporter)
 }
 
 // RenderConfig provides various ways to render a CloudConfig.
@@ -417,6 +421,12 @@ type HostnameConfig interface {
 type NetworkingConfig interface {
 	// AddNetworkConfig adds network config from interfaces to the container.
 	AddNetworkConfig(interfaces corenetwork.InterfaceInfos) error
+}
+
+// FileTransporter is the interface set on CloudConfig when it wants to optionally
+// deliver files by its own means.
+type FileTransporter interface {
+	SendBytes(hint string, payload []byte) string
 }
 
 // New returns a new Config with no options set.