--- a/provider/common/bootstrap.go
+++ b/provider/common/bootstrap.go
@@ -467,6 +467,20 @@ func ConfigureMachine(
 	cloudcfg.SetSystemUpdate(instanceConfig.EnableOSRefreshUpdate)
 	cloudcfg.SetSystemUpgrade(instanceConfig.EnableOSUpgrade)
 
+	sshinitConfig := sshinit.ConfigureParams{
+		Host:           "ubuntu@" + host,
+		Client:         client,
+		SSHOptions:     sshOptions,
+		Config:         cloudcfg,
+		ProgressWriter: ctx.GetStderr(),
+		Series:         instanceConfig.Series,
+	}
+
+	stdctx, cancel := context.WithCancel(context.TODO())
+	defer cancel()
+	ft := sshinit.NewFileTransporter(stdctx, sshinitConfig)
+	cloudcfg.SetFileTransporter(ft)
+
 	udata, err := cloudconfig.NewUserdataConfig(instanceConfig, cloudcfg)
 	if err != nil {
 		return err
@@ -481,17 +495,16 @@ func ConfigureMachine(
 	if err != nil {
 		return err
 	}
+
+	// Wait for the files to be sent to the machine.
+	if err := ft.Wait(); err != nil {
+		return errors.Annotate(err, "transporting files to machine")
+	}
+
 	script := shell.DumpFileOnErrorScript(instanceConfig.CloudInitOutputLog) + configScript
 	ctx.Infof("Running machine configuration script...")
 	// TODO(benhoyt) - plumb context through juju/utils/ssh?
-	return sshinit.RunConfigureScript(script, sshinit.ConfigureParams{
-		Host:           "ubuntu@" + host,
-		Client:         client,
-		SSHOptions:     sshOptions,
-		Config:         cloudcfg,
-		ProgressWriter: ctx.GetStderr(),
-		Series:         instanceConfig.Series,
-	})
+	return sshinit.RunConfigureScript(script, sshinitConfig)
 }
 
 // HostSSHOptionsFunc is a function that, given a hostname, returns