Instead of making a 55mb shell script, scp the juju binaries over. About 22s faster on my local machine.

#### Before

```
$ time juju bootstrap localhost
Creating Juju controller "localhost-localhost" on localhost/localhost
Looking for packaged Juju agent version 3.0-beta2 for amd64
No packaged binary found, preparing local Juju agent binary
To configure your system to better support LXD containers, please see: https://github.com/lxc/lxd/blob/master/doc/production-setup.md
Launching controller instance(s) on localhost/localhost...
 - juju-87bed7-0 (arch=amd64)
Installing Juju agent on bootstrap instance
Waiting for address
Attempting to connect to 10.75.71.165:22
Connected to 10.75.71.165
Running machine configuration script...
Bootstrap agent now started
Contacting Juju controller at 10.75.71.165 to verify accessibility...

Bootstrap complete, controller "localhost-localhost" is now available
Controller machines are in the "controller" model

Now you can run
	juju add-model <model-name>
to create a new model to deploy workloads.
juju bootstrap localhost  7.14s user 1.84s system 7% cpu 1:52.42 total
```

#### After

```
$ time juju bootstrap localhost
Creating Juju controller "localhost-localhost" on localhost/localhost
Looking for packaged Juju agent version 3.0-beta2 for amd64
No packaged binary found, preparing local Juju agent binary
To configure your system to better support LXD containers, please see: https://github.com/lxc/lxd/blob/master/doc/production-setup.md
Launching controller instance(s) on localhost/localhost...
 - juju-f2cb23-0 (arch=amd64)
Installing Juju agent on bootstrap instance
Waiting for address
Attempting to connect to 10.75.71.50:22
Connected to 10.75.71.50
Running machine configuration script...
Bootstrap agent now started
Contacting Juju controller at 10.75.71.50 to verify accessibility...

Bootstrap complete, controller "localhost-localhost" is now available
Controller machines are in the "controller" model

Now you can run
	juju add-model <model-name>
to create a new model to deploy workloads.
juju bootstrap localhost  5.04s user 1.07s system 6% cpu 1:30.33 total
```

## QA steps

bootstrap lxd/aws/etc.. (not k8s)

## Documentation changes

N/A

## Bug reference

N/A