@juanmanuel-tirado 2022-07-06
It would be interesting to see the impact of compression using gzip.