@hpidcock 2022-07-07
> It would be interesting to see the impact of compression using gzip.

Pretty much all the files we send over are already compressed, so enabling compression on the scp command will have very little impact and just add extra overhead.